test: Morph integration tests into unit tests to reduce execution time
