--- testing/camunda-process-test-java/src/test/java/io/camunda/process/test/api/SemanticSimilarityAssertBootstrapIT.java
@@ -1,32 +1,0 @@
-/*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.camunda.process.test.api;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.junit.jupiter.api.Test;
-
-@CamundaProcessTest
-public class SemanticSimilarityAssertBootstrapIT {
-
-  @Test
-  void shouldAutoBootstrapSemanticSimilarityConfigViaSpiWhenPropertiesAreSet() {
-    assertThat(CamundaAssert.getSemanticSimilarityConfig())
-        .as("SemanticSimilarityConfig should be auto-bootstrapped via SPI")
-        .isNotNull()
-        .satisfies(config -> assertThat(config.getThreshold()).isEqualTo(0.9));
-  }
-}
